Uploaded the basic flow of the working website, as it was before the Lisbon conference
Updated the styling of pages, such as the login, user-space, report.
Fixed a bug related to opening a file in PHP.  The user dashboard has
been completely rennovated, now holding 6 buttons - two of which are
active:  report and side-effects questionnaire; whilst the rest display
a popup of "coming soon...". The report page contains a basic popup of
"meetme", although all "schedule an appointment" buttons lead to the
same popup, rather than to different ones based on the practitioner of
choice.